LayoutTests/platform/chromium-win/svg/dom/path-pointAtLength-expected.txt

This tests getPointAtLength of SVG path.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pointAtLengthOfPath('M0,20 L400,20 L640,20') is '(640, 20)'
FAIL pointAtLengthOfPath('M0,20 L400,20 L640,20 z') should be (580, 19.99991226196289). Was (580, 20.000051498413086).
FAIL pointAtLengthOfPath('M0,20 L400,20 z M 320,20 L640,20') should be (100, 19.999984741210938). Was (100, 20.000009536743164).
PASS successfullyParsed is true

TEST COMPLETE
